requirements.txt
langchain
groq
python-dotenv
pypdf
neo4j
streamlit
pandas
scikit-learn
opencv-python  # For video analysis
numpy
tensorflow # If needed for ML tasks
langchain_groq
langchain_community
langchain_huggingface
langchain_chroma
docx2txt
torch
openai-whisper
ffmpeg-python